refactor: added jsconfig file to set base url
