+++ src/Components/CardList.jsx
@@ -1,4 +1,4 @@
-import Card from './Card';
+import Card from 'components/Card';
 
 function CardList({ links = [] }) {
   return (
