+++ src/styles/components/_Banner.scss
