+++ src/styles/components/_Card.scss
