+++ src/styles/components/_Footer.scss
