+++ src/styles/components/_Navigator.scss
