+++ src/styles/components/_SearchBar.scss
