+++ src/styles/components/_index.scss
